+++ src/dados/EnderecoFisico.java
@@ -1,7 +1,3 @@
-/*
- * To change this template, choose Tools | Templates
- * and open the template in the editor.
- */
 package dados;
 
 /**
@@ -12,8 +8,11 @@ package dados;
 // linear ou divido em dois endereços: página e deslocamento (offset)
 //dentro da página.
 
+
+// <<union>>
+
 public class EnderecoFisico {
-     public long end32bit; //Interpretação de um endereço virtual como um inteiro de 32 bits.
+     public int end32bit; //Interpretação de um endereço virtual como um inteiro de 32 bits.
      public Endereco endereco;//Interpretação de um endereço virtual separado em
                                 //página (MSB e LSB) e endereço
                                 //(deslocamento) dentro da página.
